fixed environment variables, removed namespace
your environment variables were a bit mixed up. you were using env variables for
one container in the other and vice versa. the docs at
https://github.com/linuxserver/docker-bookstack and https://hub.docker.com/_/mysql
list the ones to use for each. it was only the fact that the username and password for sql
had not been set properly that was causing the error, the service was set up correctly.

I also had to tell Bookstack that it was to use the url bookstack.test.com using the
APP_URL env var. This is so it could correctly redirect traffic and execute links.

even though I can now connect to the app via bookstack.test.com it doesn't seem to
display correctly. I'm unsure why this is, something to investigate. There may be a subtle issue with the ingress.